1) Exercise your assembly skills.
To write an assembly program called program_0. for the
MIPS64 architecture and to execute it.
The program must:
1. Given 2 statically initialized arrays (a and b), compute their sum and store
each result in a third array (i.e., c[i] = a[i] + b[i]). Each array
contains 50 8-bit integer numbers.
2. Search for both the maximum and minimum in the array c. The program
saves the obtained value in two variables allocated in memory, called max
and min respectively.

2) Exercise your assembly skills.

To write an assembly program called program_1.s for the
MIPS64 architecture and to execute it.
The program must:
1. Given one array of 15 8-bit integer numbers (v1), check if the content
corresponds to a palindrome sequence of numbers. If yes, use a 8-bit
unsigned variable (flag) to store the result. The variable will be equal to 1
is the sequence is palindrome, 0 otherwise.
Example of a palindrome sequence:
v1: . byte 2, 6, -3, 11, 9, 11, -3, 6, 2