1) Write an assembly program (program_1.s) for the winMIPS64 architecture described before able to implement the following piece of code described at high-level: for (i = 0; i < 60; i++){ 	v5[i] = ((v1[i]+v2[i]) * v3[i])+v4[i]; 	v6[i] = v5[i]/(v4[i]*v1[i]); v7[i] = v6[i]*(v2[i]+v3[i]); }
Assume that the vectors v1[], v2[], v3[], and v4[] are allocated previously in memory and contain 60 double precision floating point values; assume also that (v4[i]*v1[i]) does not contain 0 values. Additionally, the vectors v5[], v6[], v7[] are empty vectors also allocated in memory.

2) Write an assembly program (program_2.s) for the winMIPS64 architecture able to compute the 2D Convolution between 5x5 and 3x3 matrices, and store the result in a 3x3 matrix. The 2D convolution is a frequently used operation in many fields, such as image processing or deep learning algorithms.
The inputs of the program are the following: a single 5x5 image, also known as Input Feature Map (ifmap) in deep learning models, and a 3x3 filter (also known as kernel). In a 2D Convolution, this kernel slides over the 2D input image and performs an elementwise multiplication with the part of the input it is currently on, and then it sums up the results into a single output pixel. This operation between the ifmap and the kernel produces a 3x3 output matrix, also known as output feature map (ofmap).